therapeutic-copilot/server/requirements.txt
# ─── SAATHI AI — Backend Dependencies ────────────────────────────────────────
# FastAPI core
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9       # PostgreSQL driver (prod)
aiosqlite==0.20.0             # SQLite async driver (local dev)

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.8

# HTTP client
httpx==0.27.0
aiohttp==3.9.5

# AI / Inference
together==1.2.1               # Together AI (cloud Qwen inference for dev)
openai==1.30.5                # Compatible client for Together AI API

# RAG / Embeddings
pinecone-client==3.2.2
sentence-transformers==3.0.1

# Payments
razorpay==1.4.1

# Email / Calendar
sendgrid==6.11.0
google-api-python-client==2.131.0
google-auth-oauthlib==1.2.0

# Redis / Caching
redis==5.0.4
aioredis==2.0.1

# Utilities
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
loguru==0.7.2

# WebSockets
websockets==12.0

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
